+++ include/powsybl/iidm/TieLine.hpp
@@ -8,6 +8,7 @@
 #ifndef POWSYBL_IIDM_TIELINE_HPP
 #define POWSYBL_IIDM_TIELINE_HPP
 
+#include <powsybl/iidm/HalfLine.hpp>
 #include <powsybl/iidm/Line.hpp>
 #include <powsybl/stdcxx/math.hpp>
 
@@ -17,79 +18,7 @@ namespace iidm {
 
 class TieLine : public Line {
 public:
-    class HalfLine : public Validable {
-    public: // Validable
-        std::string getMessageHeader() const override;
-
-    public:
-        double getB1() const;
-
-        double getB2() const;
-
-        double getG1() const;
-
-        double getG2() const;
-
-        const std::string& getId() const;
-
-        const std::string& getName() const;
-
-        double getR() const;
-
-        double getX() const;
-
-        double getXnodeP() const;
-
-        double getXnodeQ() const;
-
-        bool isFictitious() const;
-
-        HalfLine& setB1(double b1);
-
-        HalfLine& setB2(double b2);
-
-        HalfLine& setFictitious(bool fictitious);
-
-        HalfLine& setG1(double g1);
-
-        HalfLine& setG2(double g2);
-
-        HalfLine& setR(double r);
-
-        HalfLine& setX(double x);
-
-        HalfLine& setXnodeP(double xnodeP);
-
-        HalfLine& setXnodeQ(double xnodeQ);
-
-    private:
-        HalfLine();
-
-        void setId(const std::string& id);
-
-        void setName(const std::string& name);
-
-        void setParent(TieLine& parent);
-
-        friend class TieLine;
-
-        friend class TieLineAdder;
-
-    private:
-        stdcxx::Reference<TieLine> m_parent;
-
-        std::string m_id;
-
-        std::string m_name;
-
-        LineCharacteristics m_lineCharacteristics;
-
-        double m_xnodeP = stdcxx::nan();
-
-        double m_xnodeQ = stdcxx::nan();
-
-        bool m_fictitious = false;
-    };
+    using HalfLine = tie_line::HalfLine;
 
 public: // Line
     double getB1() const override;
